.gitignore updated with relevant no vcs sections.
app.js was updated to include the app_api use route and require.

index.js was created for app_api to handle the route definitions and what method calls route to which javascript functions for returns.

locations.js and review.js are defined as controllers and have blank functions in them with intention of being written later.